Isolamento de validações e CSS da página "CadastroUsuario"
