+++ src/pages/CadastroUsuario/index.js
@@ -1,0 +1,3 @@
+import CadastroUsuario from "./CadastroUsuario";
+
+export default CadastroUsuario;
